# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Rocket-api",
    "description": "API Documentation for Rocket API",
    "version": "1.0.0"
  },
  "host": "myrocketapp-34740598b7f1.herokuapp.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register a new user",
        "description": "Routes for user authentication",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "exampleuser"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                }
              },
              "required": [
                "username",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully registered user and returned token",
            "schema": {
              "$ref": "#/definitions/Token"
            }
          },
          "400": {
            "description": "Invalid request body or user already exists",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login a user",
        "description": "Routes for user authentication",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "exampleuser"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                }
              },
              "required": [
                "username",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully logged in user and returned token",
            "schema": {
              "$ref": "#/definitions/Token"
            }
          },
          "400": {
            "description": "Invalid credentials or request body",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/api/landingPads/": {
      "get": {
        "tags": [
          "Landing Pads"
        ],
        "summary": "Get all landing pads",
        "description": "Routes for managing landing pads",
        "parameters": [
          {
            "name": "x-auth-token",
            "description": "Authorization token",
            "required": true,
            "type": "string",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved landing pads",
            "schema": {
              "$ref": "#/definitions/LandingPad"
            }
          },
          "400": {
            "description": "Error fetching landing pads",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Landing Pads"
        ],
        "summary": "Create a new landing pad",
        "description": "Routes for managing landing pads",
        "parameters": [
          {
            "name": "x-auth-token",
            "description": "Authorization token",
            "required": true,
            "type": "string",
            "in": "header"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "LC-39A"
                },
                "status": {
                  "type": "string",
                  "example": "active"
                },
                "location": {
                  "type": "string",
                  "example": "Cape Canaveral, FL, USA"
                }
              },
              "required": [
                "name",
                "status",
                "location"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created landing pad",
            "schema": {
              "$ref": "#/definitions/LandingPad"
            }
          },
          "400": {
            "description": "Invalid request body",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/api/landingPads/{id}": {
      "get": {
        "tags": [
          "Landing Pads"
        ],
        "summary": "Get a landing pad by ID",
        "description": "Routes for managing landing pads",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Landing pad ID"
          },
          {
            "name": "x-auth-token",
            "description": "Authorization token",
            "required": true,
            "type": "string",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved landing pad",
            "schema": {
              "$ref": "#/definitions/LandingPad"
            }
          },
          "400": {
            "description": "Invalid landing pad ID",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Landing pad not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Landing Pads"
        ],
        "summary": "Update a landing pad by ID",
        "description": "Routes for managing landing pads",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Landing pad ID"
          },
          {
            "name": "x-auth-token",
            "description": "Authorization token",
            "required": true,
            "type": "string",
            "in": "header"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "rocketId": {
                  "example": "any"
                },
                "location": {
                  "example": "any"
                },
                "surfaceType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated landing pad",
            "schema": {
              "$ref": "#/definitions/LandingPad"
            }
          },
          "400": {
            "description": "Invalid landing pad ID or request body",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Landing pad not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Landing Pads"
        ],
        "summary": "Delete a landing pad by ID",
        "description": "Routes for managing landing pads",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Landing pad ID"
          },
          {
            "name": "x-auth-token",
            "description": "Authorization token",
            "required": true,
            "type": "string",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted landing pad",
            "schema": {
              "$ref": "#/definitions/LandingPad"
            }
          },
          "400": {
            "description": "Invalid landing pad ID",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Landing pad not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Token": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string",
          "example": "string"
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string",
          "example": "string"
        }
      }
    },
...
    "LandingPad": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "string"
        },
        "rocketId": {
          "type": "string",
          "example": "ObjectId"
        },
        "location": {
          "type": "string",
          "example": "string"
        },
        "surfaceType": {
          "type": "string",
          "example": "string"
        }
      }
    },
...
  },
...
}